+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### [0.0.8]
 Fixed long press to drag the item move faster than finger.
 Fixed deleted item error.
+Add the demo gif.
 
 ### [0.0.7]