+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### [0.1.0]
+
+Updated README.md
+
 ### [0.0.9]
 
 Add onBeforeDelete event.